+++ src/main/java/com/bdaim/batch/service/BatchListService.java
@@ -57,7 +57,7 @@ public interface BatchListService {
 
     void saveBatchDetailList(List<BatchDetail> batchDetailList, String channelall, String resourceId, int certifyType, String batchId, Long operUserId, String operName);
 
-    void saveBatch(String batchname, int uploadNum, String repairStrategy, String compId, String batchId, int certifyType, String channl,String province,String city,int extNumber) throws Exception;
+    void saveBatch(String batchname, int uploadNum, String repairStrategy, String compId, String batchId, int certifyType, String channl,String province,String city,double extNumber) throws Exception;
 
     void cucIsreceive(String batchId, int cucIsReceived) throws Exception;
